Fix new message notification logic
Previously, the conversation was fetched and passed to `onNewMessage`
before updating it with the new data. This could lead to situations like
notifications being triggered for conversations that were muted since
the last message, since the `isMuted` check would still operate on the
old data.

Additionally, the first message of a new conversation did previously not
trigger a notification at all.