Fix group update
When a group update arrives, the fields are copied over the old field
values. This is a problem if the updated group leaves out optional
fields that have a default value that is different to the current value.

To fix this, the defaults always need to be set explicitly.

Fixes #563