Refactor message acknowledgements (and error handling upon reception)
With the upcoming ARP change, the Web Client will request that most
messages must be acknowledged after they have been *processed* by the
app. This is backwards compatible to the previous temporaryId stuff
which is now deprecated.

While comparing with the ARP documentation, a bunch of unhandled or
incorrectly handled error codes have been found.