Fix updating of message caption
The caption may change for example when receiving an image message on
Android. Because the internet connection may be slow, a create/message
will be sent initially (before downloading the image from the blob
server) without thumbnail and without caption (because the caption is
contained in the image EXIF data). Once the image is downloaded and
processed, an update will be sent with the caption.