Unbreak source bundle on Firefox 52
Firefox <55 has a bug that prevents the app bundle from loading:

https://stackoverflow.com/q/43307607

Unfortunately this means that only the background image is visible, and
that the version support message is not being shown at all.

The simplest workaround for now is to target a Firefox version without
native class support, to cause ES5 to be generated. Once Firefox 52 has
died out, we can remove support. This will probably (hopefully) be the
case by the end of 2019.

Note that this does not mean that we officially support Firefox 52.