Fix welcome page displaying an empty page on unlock/waiting state
Before this fix, we have interpreted a 'new-responder' event as an
indication that the connection is about to be established. However, that
is not the case since a dead responder connection can linger on the
WebSocket path for a while.

It would be better if the SaltyRTC library would signal end and start
state events which would indicate that e.g. a server or peer handshake
is in progress (has started) or has been completed (has ended). Until
we have that, we will have to live with this workaround.