Revert clearing of search when clicking on conversation / contact
It seems that when searching through a lot of conversations (I have over
230 conversations), after clicking on a conversation, the search clears
almost instantly, but it sometimes takes multiple seconds until the
conversation has opened.

This has a lot to do with the inefficient architecture of AngularJS,
where every UI update can trigger a lot of processing code, which in
turn may cause yet another UI update.

For now, we need to disable the clearing/hiding of the search when
clicking on a search result.