Fix logic error in marking messages as read (#1104)
In some  cases, a read receipt could be triggered for an outgoing
message. This wasn't an issue on Android, but on iOS this caused the
previous messages not to be marked as unread.

This commit refactors and simplifies the logic. Now read receipt should
only be sent for incoming unread non-status messages.